Moving from `types` to `typedef` due to overloaded module name
