+++ src/leapfrogai_api/typedef/request/__init__.py
@@ -1,8 +1,8 @@
-from auth_types import (
+from .auth_types import (
     CreateAPIKeyRequest as CreateAPIKeyRequest,
     ModifyAPIKeyRequest as ModifyAPIKeyRequest,
 )
-from run_create_params_base import (
+from .run_create_params_base import (
     RunCreateParamsRequestBase as RunCreateParamsRequestBase,
 )
-from run_create_params import RunCreateParamsRequest as RunCreateParamsRequest
+from .run_create_params import RunCreateParamsRequest as RunCreateParamsRequest
