+++ src/leapfrogai_api/utils/config.py
@@ -6,7 +6,7 @@ import toml
 import yaml
 from watchfiles import Change, awatch
 
-from leapfrogai_api.types import Model
+from leapfrogai_api.typedef import Model
 
 
 logger = logging.getLogger(__name__)
